[refactor] : change name BloodSellProcess -> BloodSellProcessManager and refactor
